% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn-methods.R
\name{learnSignatures}
\alias{learnSignatures}
\title{Method to Learn Signatures from SummarizedExperiment}
\usage{
learnSignatures(se, assay.type = "counts",
  method = c("PositiveProportionDifference", "PPD"), ...)
}
\arguments{
\item{se}{An object of class inheriting from "\code{\link{SummarizedExperiment}}".}

\item{assay.type}{A string specifying which assay values to use, e.g., \code{"counts"} or \code{"logcounts"}.}

\item{method}{Learning method. See section "Learning methods".}

\item{...}{Additional arguments affecting the learning method.}
}
\value{
A \code{\link{Sets}} object.
}
\description{
These method signatures learn gene set signatures optionally augmented with
(semi-)quantitative information for the prediction of sample and cell identities
in \code{SummarizedExperiment} objects.
}
\section{Learning methods}{

\describe{
\item{PositiveProportionDifference, PPD}{
\emph{Requires prior cluster membership information.}
This method computes the proportion of samples positive for each feature in each cluster,
and subsequently identifies for each cluster the features showing the maximal difference
between the detection rate in the cluster of interest and the detection rate in all other clusters.}
}
}

\examples{
# Example data ----
library(SummarizedExperiment)
nsamples <- 100
u <- matrix(rpois(20000, 2), ncol=nsamples)
rownames(u) <- paste0("Gene", sprintf("\%03d", seq_len(nrow(u))))
colnames(u) <- paste0("Cell", sprintf("\%03d", seq_len(ncol(u))))
se <- SummarizedExperiment(assays=list(counts=u))

# Example usage ----
se1 <- se
colData(se1)[, "cluster"] <- factor(sample(head(LETTERS, 3), ncol(se1), replace=TRUE))
gs <- learnSignatures(se1, method="PositiveProportionDifference", cluster.col="cluster")

relations(gs)
}
\seealso{
\code{\link[=learnMarkersByPositiveProportionDifference]{learnMarkersByPositiveProportionDifference()}}
}
\author{
Kevin Rue-Albrecht
}
