% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerDetection-methods.R
\name{makeMarkerDetectionMatrix}
\alias{makeMarkerDetectionMatrix}
\alias{makeSignatureDetectionMatrix}
\title{Identify Markers and Signatures Present in Individual Samples}
\usage{
makeMarkerDetectionMatrix(se, markers, threshold = 0,
  assay.type = "counts")

makeSignatureDetectionMatrix(matrix, object)
}
\arguments{
\item{se}{An object of class inheriting from "\code{\link{SummarizedExperiment}}".}

\item{markers}{A character vector, subset of \code{rownames(se)}.}

\item{threshold}{Value \emph{above which} the marker is considered detected.}

\item{assay.type}{A string specifying which assay values to use, e.g., "\code{counts}" or "\code{logcounts}".}

\item{matrix}{A logical matrix indicating the presence of each marker in each sample.
See \code{\link{makeMarkerDetectionMatrix}}}

\item{object}{A collection of signatures inheriting from "\code{\link{GeneSetCollection}}" or "\code{\link{tbl_geneset}}".}
}
\value{
\describe{
\item{\code{makeMarkerDetectionMatrix}}{A logical matrix indicating the presence of each marker in each sample.}
\item{\code{makeSignatureDetectionMatrix}}{A logical matrix indicating the presence of each signature in each sample.}
}
}
\description{
Create matrices of detection events for individual gene features or combination thereof (i.e., signatures).
}
\details{
The \code{makeMarkerDetectionMatrix} function declares a feature as detected if it is detected above a given threshold in a specific assay (e.g., count or UMI matrix).

The \code{makeSignatureDetectionMatrix} function declares a signature (composed of one or more gene features) as detected if all the associated features are detected.
}
\examples{
# Example data ----
library(SummarizedExperiment)
nsamples <- 100
u <- matrix(rpois(20000, 2), ncol=nsamples)
rownames(u) <- paste0("Gene", sprintf("\%03d", seq_len(nrow(u))))
se <- SummarizedExperiment(assays=list(counts=u))

gsc <- GeneSetCollection(list(
  GeneSet(setName="Cell type 1", c("Gene001", "Gene002")),
  GeneSet(setName="Cell type 2", c("Gene003", "Gene004"))
))

# Example usage ----
markerMatrix <- makeMarkerDetectionMatrix(se, unique(unlist(geneIds(gsc))))
signatureMatrix <- makeSignatureDetectionMatrix(markerMatrix, gsc)
}
\author{
Kevin Rue-Albrecht
}
