% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/markerDetection-methods.R
\docType{methods}
\name{positiveForMarker}
\alias{positiveForMarker}
\alias{positiveForMarker,SummarizedExperiment-method}
\title{Detect Samples Positive for Given Markers}
\usage{
positiveForMarker(x, row, threshold = 0, ...)

\S4method{positiveForMarker}{SummarizedExperiment}(x, row, threshold = 0,
  assay = "counts")
}
\arguments{
\item{x}{A \code{\link{SummarizedExperiment}}.}

\item{row}{Row index of markers to examine.}

\item{threshold}{Value \emph{above which} markers are considered detected.}

\item{...}{Additional arguments passed to and from methods.}

\item{assay}{Name of the \code{assay} slot to use.}
}
\value{
A logical vector (single marker) or matrix (multiple markers) where \code{TRUE} indicates detection of the marker.
}
\description{
Declare whether individual gene features (i.e., "markers") are detected above a given threshold in a given assay in each sample.
}
\examples{
# Example data ----
library(SummarizedExperiment)
ncells <- 100
u <- matrix(rpois(20000, 2), ncol=ncells)
sce <- SummarizedExperiment(assays=list(counts=u))

# Example usage ----
positiveForMarker(sce, 1, 0)
}
