% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-methods.R
\name{predict.GeneSetCollection}
\alias{predict.GeneSetCollection}
\alias{predict}
\alias{predict.tbl_geneset}
\title{Predict Method for Collection of Gene Set Signatures}
\usage{
\method{predict}{GeneSetCollection}(object, se, assay.type = "counts",
  method = c("ProportionPositive"), ...)

\method{predict}{tbl_geneset}(object, se, assay.type = "counts",
  method = c("ProportionPositive"), ...)
}
\arguments{
\item{object}{A set of signatures of class inheriting from "\code{\link{GeneSetCollection}}" or "\code{\link{tbl_geneset}}".}

\item{se}{An object of class inheriting from "\code{\link{SummarizedExperiment}}".}

\item{assay.type}{A string specifying which assay values to use, e.g., "\code{counts}" or "\code{logcounts}".}

\item{method}{Prediction method. See section "Prediction methods".}

\item{...}{Additional arguments affecting the predictions produced.}
}
\value{
The object \code{se}, updated as follows:
\itemize{
\item in the \code{metadata} slot, a \code{"Hancock"} item is added (or updated) with information tracing the prediction process (e.g., method, signatures).
\item in the \code{colData} slot, a \code{DataFrame} is nested in a new (or updated) \code{"Hancock"} column.
This DataFrame contains predicted labels in the first column and additional information in further columns for each column in \code{se}.
}
}
\description{
These method signatures apply gene set signatures optionally augmented with
(semi-)quantitative information to the prediction of sample and cell identities
in \code{SummarizedExperiment} objects.
}
\section{Prediction methods}{

\describe{
\item{ProportionPositive}{
\emph{Requires prior cluster membership information.}
Computes the proportion cells positive for each signature in each cluster.
Assigns to each cluster the signature detected in the highest proportion of cells.}
}
}

\examples{
# Example data ----
library(SummarizedExperiment)
ncells <- 100
u <- matrix(rpois(20000, 2), ncol=ncells)
rownames(u) <- paste0("Gene", sprintf("\%03d", seq_len(nrow(u))))
colnames(u) <- paste0("Cell", sprintf("\%03d", seq_len(ncol(u))))
se <- SummarizedExperiment(assays=list(counts=u))

gsc <- GeneSetCollection(list(
    GeneSet(setName="Cell type 1", c("Gene001", "Gene002")),
    GeneSet(setName="Cell type 2", c("Gene003", "Gene004"))
))

# Example usage ----
se1 <- se
colData(se1)[, "cluster"] <- factor(sample(head(LETTERS, 3), ncol(se1), replace = TRUE))
predict(gsc, se1, method="ProportionPositive", cluster.col="cluster")
}
\author{
Kevin Rue-Albrecht
}
