% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R, R/predict-methods.R
\name{barplotPredictionCount}
\alias{barplotPredictionCount}
\alias{barplotPredictionProportion}
\alias{reducedDimPrediction}
\alias{predict-hancock}
\alias{predict.GeneSetCollection}
\alias{predict}
\alias{predict-methods}
\alias{predict.tbl_geneset}
\alias{predict.BaseSets}
\title{Predict Method for Collection of Gene Set Signatures}
\usage{
barplotPredictionCount(se, highlight = character(0), labels = TRUE)

barplotPredictionProportion(se, highlight = character(0),
  labels = TRUE)

reducedDimPrediction(se, highlight = character(0), redDimType = "PCA",
  x = 1, y = 2, labels = TRUE)

\method{predict}{GeneSetCollection}(object, se, assay.type = "counts",
  method = c("ProportionPositive", "PP"), ...)

\method{predict}{tbl_geneset}(object, se, assay.type = "counts",
  method = c("ProportionPositive", "PP"), ...)

\method{predict}{BaseSets}(object, se, assay.type = "counts",
  method = c("ProportionPositive", "PP"), ...)
}
\arguments{
\item{se}{An object of class inheriting from \code{\link[=RangedSummarizedExperiment-class]{SummarizedExperiment}}.}

\item{highlight}{Character vector indicating names of signatures to highlight.}

\item{labels}{Logical value indicating whether to display labels.}

\item{redDimType}{Name of the reduced dimension result type to display.}

\item{x}{Name of the covariate to display on the x-axis.}

\item{y}{Name of the covariate to display on the y-axis.}

\item{object}{A set of signatures of class inheriting from \code{\linkS4class{BaseSets}} or \code{\linkS4class{GeneSetCollection}}.}

\item{assay.type}{A string specifying which assay values to use, e.g., \code{"counts"} or \code{"logcounts"}.}

\item{method}{Prediction method. See section "Prediction methods".}

\item{...}{Additional arguments affecting the predictions produced.}
}
\value{
The object \code{se}, updated as follows:
\itemize{
\item in the \code{metadata} slot, a \code{"hancock"} item is added (or updated) with information tracing the prediction process (e.g., method, signatures).
\item in the \code{"colData"} slot, a \code{DataFrame} is nested in a new (or updated) \code{"hancock"} column.
This DataFrame contains predicted labels in the first column and additional information in further columns for each column in \code{se}.
}
}
\description{
These method signatures apply gene set signatures optionally augmented with
(semi-)quantitative information to the prediction of sample and cell identities
in \code{\link[=RangedSummarizedExperiment-class]{SummarizedExperiment}} objects.
}
\section{Functions}{
\itemize{
\item \code{barplotPredictionCount}: Returns a \code{ggplot} bar plot displaying
the count of samples predicted for each gene signature.

\item \code{barplotPredictionProportion}: Returns a \code{ggplot} bar plot displaying
the proportion of samples predicted for each gene signature.

\item \code{reducedDimPrediction}: Returns a \code{ggplot} scatter plot displaying
the first reduced dimension result in \code{reducedDims(se)}.
}}

\section{Prediction methods}{

\describe{
\item{ProportionPositive, PP}{
\emph{Requires prior cluster membership information.}
Computes the proportion of samples positive for each signature in each cluster.
Assigns to each cluster the signature detected in the highest proportion of samples.}
}
}

\examples{
# Example data ----
library(SummarizedExperiment)
nsamples <- 100
u <- matrix(rpois(20000, 2), ncol=nsamples)
rownames(u) <- paste0("Gene", sprintf("\%03d", seq_len(nrow(u))))
colnames(u) <- paste0("Cell", sprintf("\%03d", seq_len(ncol(u))))
se <- SummarizedExperiment(assays=list(counts=u))

bs <- BaseSets(
    relations=DataFrame(
        element = c("Gene001", "Gene002", "Gene003", "Gene004"),
        set     = c(rep("Cell type 1", 2), rep("Cell type 2", 2))
    )
)

# Example usage ----

se1 <- se
colData(se1)[, "cluster"] <- factor(sample(head(LETTERS, 3), ncol(se1), replace=TRUE))
se1 <- predict(bs, se1, method="ProportionPositive", cluster.col="cluster")
# Visualise the count of samples predicted for each signature in each cluster
barplotPredictionCount(se1, highlight=c("Cell type 1"))

barplotPredictionProportion(se1, highlight=c("Cell type 2"))

library(SingleCellExperiment)
sce1 <- as(se1, "SingleCellExperiment")
reducedDim(sce1, "PCA") <- prcomp(t(assay(sce1)))$x
reducedDimPrediction(sce1, highlight="Cell type 1", redDimType="PCA", x=1, y=2)
}
\seealso{
\code{\link{predictByProportionPositive}}.
}
\author{
Kevin Rue-Albrecht
}
