% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyLabels.R
\name{shinyLabels}
\alias{shinyLabels}
\title{Interactively Inspect and Name Gene Signatures}
\usage{
shinyLabels(gs, se)
}
\arguments{
\item{gs}{A set of gene signatures inheriting from \code{\link{BaseSets}}.}

\item{se}{An object of class inheriting from \code{\link[=RangedSummarizedExperiment-class]{SummarizedExperiment}}.}
}
\value{
The updated set of gene signatures as a \code{\link{BaseSets}}.
}
\description{
This function launches a Shiny app to inspect the gene signatures defined in a \code{BaseSets}
for the purpose of (re-)naming those signatures interactively.
The app returns the updated \code{BaseSets} when closed using the "Done" button.
}
\examples{
# Example data ----

bs <- BaseSets(
    relations=DataFrame(
        element = c("Gene001", "Gene002", "Gene003", "Gene004"),
        set     = c(rep("Cell type 1", 2), rep("Cell type 2", 2))
    )
)

library(SummarizedExperiment)
nsamples <- 100
u <- matrix(rpois(20000, 2), ncol=nsamples)
rownames(u) <- paste0("Gene", sprintf("\%03d", seq_len(nrow(u))))
colnames(u) <- paste0("Cell", sprintf("\%03d", seq_len(ncol(u))))
se <- SummarizedExperiment(assays=list(counts=u))
colData(se)[, "cluster"] <- factor(sample(head(LETTERS, 3), ncol(se), replace=TRUE))

se1 <- predict(bs, se, method="ProportionPositive", cluster.col="cluster")

# Add reduced dimension results to enable app features
library(SingleCellExperiment)
library(Rtsne)
sce1 <- as(se1, "SingleCellExperiment")
reducedDim(sce1, "PCA") <- prcomp(t(assay(sce1)))$x
reducedDim(sce1, "TSNE") <- Rtsne(X=reducedDim(sce1, "PCA"))$Y

# Example usage ----

if (interactive()){
    library(shiny)
    x <- runApp(shinyLabels(bs, sce1))
}
}
