% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/GeneSet-helpers.R,
%   R/GeneSetCollection-helpers.R
\docType{methods}
\name{uniqueMarkers}
\alias{uniqueMarkers}
\alias{uniqueMarkers,tbl_geneset-method}
\alias{uniqueMarkers,GeneSetCollection-method}
\title{Extract Unique Markers from a Collection of Gene Sets}
\usage{
uniqueMarkers(object)

\S4method{uniqueMarkers}{tbl_geneset}(object)

\S4method{uniqueMarkers}{GeneSetCollection}(object)
}
\arguments{
\item{object}{A \code{\link{tbl_geneset}} or \code{GeneSetCollection}.}
}
\value{
A character vector of unique markers across all gene sets.
}
\description{
Extract Unique Markers from a Collection of Gene Sets
}
\examples{
# Example data ----
library(GeneSet)
tgs <- tbl_geneset(
    "Cell type 1"=c("Gene001", "Gene002"),
    "Cell type 2"=c("Gene002", "Gene003", "Gene004")
)

# Example usage ----
uniqueMarkers(tgs)
}
\author{
Kevin Rue-Albrecht
}
