% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hancock-package.R
\docType{package}
\name{hancock-package}
\alias{hancock}
\alias{hancock-package}
\title{hancock: Learn and Apply Cell Type Signatures}
\description{
Applies transcriptional signatures to assign cell identities.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kevinrue/hancock}
  \item \url{http://kevinrue.github.io/hancock}
  \item Report bugs at \url{https://github.com/kevinrue/hancock/issues}
}

}
\author{
\strong{Maintainer}: Kevin Rue-Albrecht \email{kevin.rue-albrecht@kennedy.ox.ac.uk} (0000-0003-3899-3872)

Authors:
\itemize{
  \item Robert Amezquita
  \item Aaron Lun \email{infinite.monkeys.with.keyboards@gmail.com}
}

}
\keyword{internal}
