% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hancock-package.R
\docType{package}
\name{Hancock-package}
\alias{Hancock}
\alias{Hancock-package}
\title{Hancock: Learn and Apply Cell Type Signatures}
\description{
Applies transcriptional signatures to assign cell identities.
}
\author{
\strong{Maintainer}: Kevin Rue-Albrecht \email{kevin.rue-albrecht@kennedy.ox.ac.uk}

Authors:
\itemize{
  \item Robert Amezquita
}

}
\keyword{internal}
