% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerDetection-methods.R
\name{makeMarkerDetectionMatrix}
\alias{makeMarkerDetectionMatrix}
\alias{makeMarkerProportionScree}
\alias{makeSignatureDetectionMatrix}
\alias{makeMarkerProportionMatrix}
\title{Identify Markers and Signatures Present in Individual Samples}
\usage{
makeMarkerDetectionMatrix(se, markers, threshold = 0,
  assay.type = "counts")

makeMarkerProportionScree(matrix)

makeSignatureDetectionMatrix(matrix, object)

makeMarkerProportionMatrix(se, cluster.col, assay.type = "counts",
  threshold = 0)
}
\arguments{
\item{se}{An object of class inheriting from \code{\link[=RangedSummarizedExperiment-class]{SummarizedExperiment}}.}

\item{markers}{A character vector, subset of \code{rownames(se)}.}

\item{threshold}{Value \emph{above which} the marker is considered detected.}

\item{assay.type}{A string specifying which assay values to use, e.g., \code{"counts"} or \code{"logcounts"}.}

\item{matrix}{A logical matrix indicating the presence of each marker (row) in each sample (column).}

\item{object}{A collection of signatures inheriting from "\code{\link{GeneSetCollection}}" or "\code{\link{BaseSets}}".}

\item{cluster.col}{Name of a column in \code{colData(se)} that contains
a factor indicating cluster membership for each column (i.e. sample) in \code{se}.}
}
\value{
\describe{
\item{\code{makeMarkerDetectionMatrix}}{A \code{logical} matrix indicating detectable levels of each marker in each sample.}
\item{\code{makeSignatureDetectionMatrix}}{A \code{logical} matrix indicating detectable levels of each signature in each sample.}
\item{\code{makeMarkerProportionMatrix}}{A matrix indicating for each feature the proportion of samples expressing detectable levels in each cluster.}
\item{\code{makeMarkerProportionScree}}{A \code{double} vector indicating the proportion of samples positive for markers in the input \code{logical} matrix.}
}
}
\description{
Create matrices of detection events for individual gene features or combination thereof (i.e., signatures).
}
\details{
The \code{makeMarkerDetectionMatrix} function returns a marker by sample matrix declaring a feature as detected if it is detected above a given threshold in a specific assay (e.g., \code{"counts"}, \code{"logcounts"}).

The \code{makeSignatureDetectionMatrix} function returns a sample by signature matrix declaring a signature (composed of one or more gene features) as detected if all the associated features are detected.

The \code{makeMarkerProportionMatrix} function returns a signature by cluster matrix indicating the proportion of samples expressing detectable levels of each marker in individual predefined clusters.

The \code{makeMarkerProportionScree} function compute the 'cumulative' (i.e., combined) detection rate of markers:
the proportion of samples with detectable levels of the first marker, both of the first two markers, etc.
}
\examples{
# Example data ----
library(SummarizedExperiment)
nsamples <- 100
u <- matrix(rpois(20000, 2), ncol=nsamples)
rownames(u) <- paste0("Gene", sprintf("\%03d", seq_len(nrow(u))))
se <- SummarizedExperiment(assays=list(counts=u))

geneLists <- list(
  "Cell type 1" = c("Gene001", "Gene002"),
  "Cell type 2" = c("Gene003", "Gene004")
)
bs <- as(geneLists, "BaseSets")

# Example usage ----

markerMatrix <- makeMarkerDetectionMatrix(se, ids(elementData(bs)))
signatureMatrix <- makeSignatureDetectionMatrix(markerMatrix, bs)

tab <- makeMarkerProportionScree(markerMatrix)
plot(tab$markers, tab$proportion, ylim=c(0, 1), main="Combined detection scree")

se$cluster <- factor(sample(c("A", "B"), ncol(se), TRUE))
proportionMatrix <- makeMarkerProportionMatrix(
    se, cluster.col="cluster", assay.type="counts", threshold=0)
head(proportionMatrix)
}
\author{
Kevin Rue-Albrecht, with C++ code by Aaron Lun
}
