% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerDetection-methods.R
\name{makeMarkerDetectionMatrix}
\alias{makeMarkerDetectionMatrix}
\alias{makeSignatureDetectionMatrix}
\title{Identify Markers and Signatures Present in Individual Samples}
\usage{
makeMarkerDetectionMatrix(se, markers, threshold = 0,
  assay.type = "counts")

makeSignatureDetectionMatrix(matrix, object)
}
\arguments{
\item{se}{An object of class inheriting from "\code{\link{SummarizedExperiment}}".}

\item{markers}{A character vector, subset of}

\item{threshold}{Value \emph{above which} the marker is considered detected.}

\item{assay.type}{A string specifying which assay values to use, e.g., "\code{counts}" or "\code{logcounts}".}

\item{matrix}{A logical matrix indicating the presence of each marker in each sample.
See \code{\link{makeMarkerDetectionMatrix}}}

\item{object}{A set of signatures of class inheriting from "\code{\link{GeneSetCollection}}".}
}
\value{
\describe{
\item{\code{makeMarkerDetectionMatrix}}{A logical matrix indicating the presence of each marker in each sample.}
\item{\code{makeSignatureDetectionMatrix}}{A logical matrix indicating the presence of each signature in each sample.}
}
}
\description{
Identify Markers and Signatures Present in Individual Samples
}
\examples{
# Example data ----
library(SummarizedExperiment)
ncells <- 100
u <- matrix(rpois(20000, 2), ncol=ncells)
rownames(u) <- paste0("Gene", sprintf("\%03d", seq_len(nrow(u))))
se <- SummarizedExperiment(assays=list(counts=u))

gsc <- GeneSetCollection(list(
  GeneSet(setName="Cell type 1", c("Gene001", "Gene002")),
  GeneSet(setName="Cell type 2", c("Gene003", "Gene004"))
))

# Example usage ----
markerMatrix <- makeMarkerDetectionMatrix(se, unique(unlist(geneIds(gsc))))
signatureMatrix <- makeSignatureDetectionMatrix(markerMatrix, gsc)
}
\author{
Kevin Rue-Albrecht
}
