% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/GeneSet-helpers.R,
%   R/GeneSetCollection-helpers.R
\docType{methods}
\name{makeFilterExpression}
\alias{makeFilterExpression}
\alias{makeFilterExpression,tbl_geneset-method}
\alias{makeFilterExpression,GeneSetCollection-method}
\title{Build Filter Expressions from GeneSetCollection objects}
\usage{
makeFilterExpression(object)

\S4method{makeFilterExpression}{tbl_geneset}(object)

\S4method{makeFilterExpression}{GeneSetCollection}(object)
}
\arguments{
\item{object}{A \code{\link{tbl_geneset}} or \code{\link{GeneSetCollection}}.}
}
\value{
A list of \code{\link{expression}} that combines the markers
listed in each gene set.
}
\description{
This function create a list of unevaluated expressions representing a collection of signatures.
The resulting expressions can be evaluated as \code{FilterRules} inside an environment such as a \code{data.frame} of sample-by-gene detection events.
}
\examples{
# Example data ----
library(GeneSet)
tgs <- tbl_geneset(
    "Cell type 1"=c("Gene001", "Gene002"),
    "Cell type 2"=c("Gene002", "Gene003", "Gene004")
)

# Example usage ----
makeFilterExpression(tgs)
}
\author{
Kevin Rue-Albrecht
}
