% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/markedDetection-methods.R
\docType{methods}
\name{positiveForMarker}
\alias{positiveForMarker}
\alias{positiveForMarker,SummarizedExperiment-method}
\title{Detect cells positive for a given marker}
\usage{
positiveForMarker(x, row, threshold = 0, ...)

\S4method{positiveForMarker}{SummarizedExperiment}(x, row, threshold = 0,
  assay = "counts")
}
\arguments{
\item{x}{A \code{\link{SummarizedExperiment}} or \code{matrix}.}

\item{row}{Row index of a marker.}

\item{threshold}{Value \emph{above which} the marker is considered detected.}

\item{...}{Additional arguments passed to and from methods.}

\item{assay}{Name of the \code{assay} slot to use.}
}
\value{
A logical vector where \code{TRUE} indicates detection of the marker.
}
\description{
Detect cells positive for a given marker
}
\examples{
# Example data ----
library(SummarizedExperiment)
ncells <- 100
u <- matrix(rpois(20000, 2), ncol=ncells)
sce <- SummarizedExperiment(assays=list(counts=u))

# Example usage ----
positiveForMarker(sce, 1, 0)
}
