% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_labels.R
\name{shinyLabels}
\alias{shinyLabels}
\title{Interactively Inspect and Name Gene Signatures}
\usage{
shinyLabels(gs)
}
\arguments{
\item{gs}{A set of gene signatures inheriting from \code{\link{tbl_geneset}}.}
}
\value{
The update set of gene signatures as a \code{\link{tbl_geneset}}.
}
\description{
This function launches a Shiny app to inspect the gene signatures defined in a \code{tbl_geneset}
for the purpose of (re-)naming those signatures interactively.
The app returns the updated \code{tbl_geneset} when closed using the \code{"Done"} button.
}
\examples{
# Example data ----
library(GeneSet)
tgs <- tbl_geneset(
    "Cell type 1"=c("Gene001", "Gene002"),
    "Cell type 2"=c("Gene002", "Gene003", "Gene004")
)

# Example usage ----
if (interactive()){
x <- runApp(shinyLabels(tgs))
}
}
