% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/BaseSets-helpers.R,
%   R/GeneSet-helpers.R, R/GeneSetCollection-helpers.R
\docType{methods}
\name{uniqueMarkerNames}
\alias{uniqueMarkerNames}
\alias{uniqueSetNames}
\alias{uniqueMarkerNames,BaseSets-method}
\alias{uniqueSetNames,BaseSets-method}
\alias{uniqueMarkerNames,tbl_geneset-method}
\alias{uniqueSetNames,tbl_geneset-method}
\alias{uniqueMarkerNames,GeneSetCollection-method}
\alias{uniqueSetNames,GeneSetCollection-method}
\title{Extract Unique Signature Names from a Collection of Gene Sets}
\usage{
uniqueMarkerNames(object)

uniqueSetNames(object)

\S4method{uniqueMarkerNames}{BaseSets}(object)

\S4method{uniqueSetNames}{BaseSets}(object)

\S4method{uniqueMarkerNames}{tbl_geneset}(object)

\S4method{uniqueSetNames}{tbl_geneset}(object)

\S4method{uniqueMarkerNames}{GeneSetCollection}(object)

\S4method{uniqueSetNames}{GeneSetCollection}(object)
}
\arguments{
\item{object}{An object of class inheriting from \code{\linkS4class{BaseSets}} or \code{\linkS4class{GeneSetCollection}}.}
}
\value{
A character vector of unique set or marker names across all gene sets.
}
\description{
\code{uniqueMarkerNames} and \code{uniqueSetNames} extract the character vectors of unique marker and set names from objects that store collections of gene sets.
}
\examples{
# Example data ----

library(unisets)
gmt <- system.file(package = "hancock", "extdata", "seurat_pbmc3k.gmt")
genesets <- import(gmt)

# Usage ----
um <- uniqueMarkerNames(genesets)
us <- uniqueSetNames(genesets)
}
\author{
Kevin Rue-Albrecht
}
